requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
psycopg2-binary==2.9.5
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.10.2
pytest==7.3.1
requests==2.28.2
python-dotenv==1.0.0
reportlab==3.6.12
